move times into Timers.cs
