+++ src/VIP-Plugin.cs
@@ -4,7 +4,6 @@ using Core.Models;
 using CounterStrikeSharp.API;
 using CounterStrikeSharp.API.Core;
 using CounterStrikeSharp.API.Modules.Memory;
-using Microsoft.Extensions.Logging;
 using static CounterStrikeSharp.API.Core.Listeners;
 
 namespace Core
@@ -74,69 +73,7 @@ namespace Core
         {
             VirtualFunctions.CBaseEntity_TakeDamageOldFunc.Unhook(OnTakeDamage, HookMode.Pre);
         }
-        public void HealthRegenCallback(object state)
-        {
-            var group = (VipGroupConfig)state;
 
-            if (group == null)
-            {
-                Logger.LogError("group is null in HealthRegenCallback");
-                return;
-            }
-
-            Server.NextFrame(() =>
-            {
-                foreach (var player in Utilities.GetPlayers())
-                {
-                    if (!player.IsValid ||
-                        !player.PawnIsAlive ||
-                        !_playerCache.ContainsKey(player))
-                    {
-                        continue;
-                    }
-
-                    var playerData = _playerCache[player];
-                    if (playerData.GroupId == -1 || Config.VIPGroups[playerData.GroupId] != group)
-                    {
-                        continue;
-                    }
-
-                    PlayerManager.AddHealth(player, group.Misc.HealthRegen.Amount, group.Limits.MaxHp);
-                }
-            });
-        }
-
-        public void ArmorRegenCallback(object state)
-        {
-            var group = (VipGroupConfig)state;
-
-            if (group == null)
-            {
-                Logger.LogError("group is null in ArmorRegenCallback");
-                return;
-            }
-
-            Server.NextFrame(() =>
-            {
-                foreach (var player in Utilities.GetPlayers())
-                {
-                    if (!player.IsValid ||
-                        !player.PawnIsAlive ||
-                        !_playerCache.ContainsKey(player))
-                    {
-                        continue;
-                    }
-
-                    var playerData = _playerCache[player];
-                    if (playerData.GroupId == -1 || Config.VIPGroups[playerData.GroupId] != group)
-                    {
-                        continue;
-                    }
-
-                    PlayerManager.AddArmor(player, group.Misc.ArmorRegen.Amount);
-                }
-            });
-        }
         private static CCSGameRules GetGamerules()
         {
             return Utilities.FindAllEntitiesByDesignerName<CCSGameRulesProxy>("cs_gamerules").First().GameRules!;
